add final features to the form
