+++ next.config.js
@@ -17,7 +17,8 @@ const nextConfig = {
     remotePatterns: [
       {
         protocol: "https",
-        hostname: "source.unsplash.com",
+        hostname: "**",
+        pathname: "**",
       },
     ],
   },
