+++ package.json
@@ -17,7 +17,8 @@
     "next": "13.1.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "styled-components": "^5.3.6"
+    "styled-components": "^5.3.6",
+    "uid": "^2.0.1"
   },
   "devDependencies": {
     "@svgr/webpack": "^6.5.1",
