+++ pages/upload.js
@@ -1,5 +1,6 @@
-import { Fragment, useState } from "react";
-import SVGIcon from "@/components/SVGIcon";
+import UploadForm from "@/components/UploadForm";
+import initialPictures from "@/store";
+import { useRouter } from "next/router";
 
 // For the search function I got inspierd by: https://www.youtube.com/watch?v=Jd7s7egjt30&ab_channel=ReactwithMasoud
 
@@ -18,174 +19,14 @@ const initialCategories = [
 ];
 
 export default function UploadPage() {
-  const [categories, setCategories] = useState(initialCategories); // a list of all possible categories depending on users selection
-  const [inputCategoryValue, setInputCategoryValue] = useState(""); // using for controlled Inputs
-  const [isCategoryFound, setIsCategoryFound] = useState(false); // for leting the user know if there is no result to the inputValue
-  const [selectedCategories, setSelectedCategories] = useState([]); // a list of all selected categories
-  const [noCategoriesSelected, setNoCategoriesSelected] = useState(false); // letting the user know that there must be atleast one category selected
+  const router = useRouter();
 
-  function onChangeInput(event) {
-    setNoCategoriesSelected(false);
-    setInputCategoryValue(event.target.value);
-    if (event.target.value < 1) {
-      setIsCategoryFound(false);
-      return;
-    }
-    setIsCategoryFound(true);
+  function handleSubmit(picture) {
+    initialPictures.push(picture);
+    console.log(initialPictures);
+    router.push("/");
   }
-
-  function handleAddSelectedCategory(selectedCategory) {
-    setSelectedCategories([...selectedCategories, selectedCategory]); // adds the category to the selected categories
-    setCategories(
-      categories.filter((category) => category !== selectedCategory) // removes the category from all categories so that I cant choose it again
-    );
-    setInputCategoryValue(""); // sets the Value of the Input to zero
-  }
-
-  function handleDeleteSelectedCategory(selectedCategory) {
-    setSelectedCategories(
-      selectedCategories.filter((category) => category !== selectedCategory) // removes the category from selected categories
-    );
-    setCategories([...categories, selectedCategory]); // adds the category to the all categories
-  }
-
-  // creates the array that shows the search results
-  const filterdCategories = categories.filter((category) => {
-    const searchTerm = inputCategoryValue.toLowerCase();
-    const categoryLower = category.toLowerCase();
-
-    return searchTerm && categoryLower.startsWith(searchTerm);
-  });
-
-  function handleSubmit(event) {
-    event.preventDefault();
-
-    const formData = new FormData(event.target);
-    const data = Object.fromEntries(formData);
-
-    const currentDate = new Date().toLocaleDateString("us-US", {
-      dateStyle: "medium",
-    });
-
-    const uploadObject = {
-      image: data.image,
-      describtion: data.describtion,
-      date: currentDate,
-      categories: selectedCategories,
-      settings: {
-        film: data.film,
-        aperture: data.aperture,
-        time: data.time,
-        lens: data.lens,
-        camera: data.camera,
-      },
-    };
-    console.log(uploadObject);
-  }
-
-  console.log("selectedCategories", selectedCategories);
-
-  // --------------------------------------------------------------------------------
   return (
-    <form onSubmit={handleSubmit}>
-      <label htmlFor="name">upload image</label>
-      <input type="text" id="image" name="image" required />
-      <br />
-      <label htmlFor="describtion">describtion</label>
-      <textarea type="text" id="describtion" name="describtion" required />
-
-      {/* ------- serach function + suggestions ------- */}
-      <fieldset>
-        <h2>Add categories</h2>
-        <SVGIcon variant="search" width="13" />
-        <input
-          type="text"
-          value={inputCategoryValue}
-          onChange={onChangeInput}
-          onBlur={() => setIsCategoryFound(false)}
-          style={{
-            outline: noCategoriesSelected && "red 2px solid",
-          }}
-        />
-        {noCategoriesSelected && <span>required</span>}
-        <div>
-          {filterdCategories.length >= 1
-            ? filterdCategories.map((category) => {
-                return (
-                  <p
-                    key={category}
-                    onClick={() => {
-                      handleAddSelectedCategory(category);
-                    }}
-                  >
-                    {category}
-                  </p>
-                );
-              })
-            : isCategoryFound && <p>no categories found</p>}
-        </div>
-
-        {/* ------- listing all selected categories ------- */}
-        <section>
-          <h3>selected categories</h3>
-          {selectedCategories < 1 ? (
-            <p>no categories selected</p>
-          ) : (
-            selectedCategories.map((category) => {
-              return (
-                <Fragment key={category}>
-                  <p>{category}</p>
-                  <button
-                    onClick={() => {
-                      handleDeleteSelectedCategory(category);
-                    }}
-                  >
-                    delete
-                  </button>
-                </Fragment>
-              );
-            })
-          )}
-        </section>
-      </fieldset>
-
-      <fieldset>
-        <h2>Settings</h2>
-        <SVGIcon variant="film" width="16" />
-        <label htmlFor="film">film</label>
-        <input type="text" id="film" name="film" />
-        <br />
-        <SVGIcon variant="aperture" width="13" />
-        <label htmlFor="aperture">aperture</label>
-        <span> f1/</span>
-        <input type="number" id="aperture" name="aperture" step=".01" />
-        <br />
-        <SVGIcon variant="time" width="13" />
-        <label htmlFor="time">time</label>
-        <input type="text" id="time" name="time" />
-        <span>s</span>
-        <br />
-        <SVGIcon variant="lens" width="13" />
-        <label htmlFor="lens">lens</label>
-        <input type="tel" id="lens" name="lens" step="any" />
-        <span>mm</span>
-        <br />
-        <SVGIcon variant="camera" width="15" />
-        <label htmlFor="camera">camera</label>
-        <input type="text" id="camera" name="camera" />
-      </fieldset>
-      {selectedCategories < 1 ? (
-        <button
-          type="button"
-          onClick={() => {
-            setNoCategoriesSelected(true);
-          }}
-        >
-          no upload
-        </button>
-      ) : (
-        <button type="submit">upload</button>
-      )}
-    </form>
+    <UploadForm initialCategories={initialCategories} onSubmit={handleSubmit} />
   );
 }
